Fix a regression in CJBig2_SDDProc::DecodeArith().
In https://pdfium-review.googlesource.com/92890, the refactoring
accidentally changed a uint32_t variable into a uint8_t variable. As a
result, some calculations in CJBig2_SDDProc are incorrect and some
JPEG2000 images fail to render.

Note that there is no test case here, because the PDF that demonstrates
the bug contains PII, and it is not immediately obvious how to create a
JPEG2000 image with the same properties.

Bug: chromium:1329994
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/93970
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fxcodec/jbig2/JBig2_SddProc.cpp
@@ -163,7 +163,7 @@ std::unique_ptr<CJBig2_SymbolDict> CJBig2_SDDProc::DecodeArith(
           if (!BS)
             return nullptr;
         } else if (REFAGGNINST == 1) {
-          uint8_t SBNUMSYMS = SDNUMINSYMS + NSYMSDECODED;
+          uint32_t SBNUMSYMS = SDNUMINSYMS + NSYMSDECODED;
           uint32_t IDI;
           IAID->Decode(pArithDecoder, &IDI);
           if (IDI >= SBNUMSYMS)
